[NayNay] Create Account Refactor + Testing
- converted creating an account to pure function
- added tests for creating an account
- renamed some files for organization and consistency
